back-end/klasifikasi_hama/requirements.txt
absl-py==2.2.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.3.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.2
docopt==0.6.2
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.12
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.10.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
namex==0.0.9
narwhals==1.29.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.1.3
oauthlib==3.2.2
openai==0.25.0
openpyxl==3.0.10
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pandas-stubs==1.5.2.221213
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pillow==11.2.1
pipreqs==0.5.0
platformdirs==4.3.7
prettytable==3.3.0
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pyarrow==19.0.1
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytweening==1.0.7
pytz==2025.1
pywhatkit==5.4
pywin32==308
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==1.3.1
rich==14.0.0
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.28.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.43.0
sympy==1.12
tabulate==0.8.9
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.2
tqdm==4.64.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
tweepy==4.14.0
types-pytz==2022.7.0.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.4.0
uvicorn==0.34.2
var-dump==1.2
watchdog==6.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
yarg==0.1.9
